refactor: Separate pick one from `ModalPickOne` into separate component so we can re-use it
