+++ src/components/ModalPickOne/component.tsx
@@ -1,11 +1,9 @@
-import React, { useState, useCallback, useMemo } from 'react';
+import React, { useMemo } from 'react';
 
-import { Modal } from '../Modal';
 import { useBuildTestId } from '../../modules/test-ids';
-import { TextInput } from '../TextInput';
-import { Space } from '../Space';
+import { PickOne } from '../internal/PickOne';
+import { Modal } from '../Modal';
 
-import { Content, Header } from './styled';
 import { ModalPickOneContext } from './context';
 
 export type Props = React.ComponentPropsWithoutRef<typeof Modal>;
@@ -13,44 +11,11 @@ export type Props = React.ComponentPropsWithoutRef<typeof Modal>;
 export const Component = ({ onClose, 'data-testid': testId, children, ...otherProps }: Props) => {
   const context = useMemo(() => ({ onClose, testId }), [onClose, testId]);
   const buildTestId = useBuildTestId(testId);
-  const [search, setSearch] = useState('');
-  const updateSearch = useCallback(
-    (evt: React.ChangeEvent<HTMLInputElement>) => setSearch(evt.target.value),
-    [],
-  );
-
-  const lowerCaseSearch = useMemo(() => search.toLowerCase(), [search]);
-
-  const filteredResults = useMemo(
-    () =>
-      React.Children.toArray(children).filter((it) => {
-        if (!React.isValidElement<{ searchAs: string[] | string }>(it)) return false;
-
-        const { searchAs } = it.props;
-        if (!searchAs) return false;
-
-        if (typeof searchAs === 'string') {
-          return searchAs.toLowerCase().includes(lowerCaseSearch);
-        }
-
-        if (Array.isArray(searchAs)) {
-          return !!searchAs.find((search) => search.toLowerCase().includes(lowerCaseSearch));
-        }
-
-        return false;
-      }),
-    [children, lowerCaseSearch],
-  );
 
   return (
     <ModalPickOneContext.Provider value={context}>
       <Modal {...otherProps} onClose={onClose} data-testid={buildTestId()}>
-        <Header>
-          <TextInput value={search} onChange={updateSearch} data-testid={buildTestId('input')} />
-        </Header>
-        <Space size="increased" />
-        <Content>{filteredResults}</Content>
-        <Space size="increased" />
+        <PickOne>{children}</PickOne>
       </Modal>
     </ModalPickOneContext.Provider>
   );
