+++ src/components/PickOne/index.tsx
@@ -1,0 +1,1 @@
+export { Component as PickOne } from './component';
